% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_cellphase.R
\name{gg_cellphase}
\alias{gg_cellphase}
\title{Create's a ggplot for dendrometer cell phase data.}
\usage{
gg_cellphase(gg_cellphase_data, start.day, end.day, date_breaks, date_labels, type = "all")
}
\arguments{
\item{gg_cellphase_data}{a \code{data.frame} with either gap-free or gap-filled dendrometer phase series as produced by \link{gg_cellphase_data}.}

\item{start.day}{a \code{timestamp} specifying the begining of dendrometer data to be used in the final plot timestamp format is (\code{\%Y-\%m-\%d \%H:\%M:\%S} format).}

\item{end.day}{a \code{timestamp} specifying the end of dendrometer data to be used in the final plot timestamp format is (\code{\%Y-\%m-\%d \%H:\%M:\%S} format).}

\item{date.breaks}{a \code{string} giving the distance between breaks such as: "2 weeks", or "10 years".}

\item{date.labels}{a \code{string} giving the formatting specification for the labels. Codes are defined in \link[base]{strftime}. If both labels and date_labels are specified, date_labels wins.}

\item{type}{a \code{string} specifying either 1) the default \code{type = "all"}; a single plot with all dendrometer data on it, or 2) \code{type = "facet"}; if the plot should be displayed based on the number of dendrometers used within the users origional data frame.}
}
\value{
The function returns a \code{geom_point} plot or can be stored as an object.
}
\description{
Can be displayed as a single plot or as a facet wrap for each unique dendrometer.
}
\examples{
\donttest{
#plot all on one
library(dendrometeR)
data(dmCD)
dm.phase <- phase_def(dmCD)
dm.cellphase <- gg_cellphase_data(dmCD)
gg_cellphase(gg_cellphase_data = dm.cellphase)

#plot facet for each dendrometer
library(dendrometeR)
data(dmED)
dmED <- fill_gaps(dmED)
dm.phase <- phase_def(dmED)
dm.cellphase <- gg_cellphase_data(dmED)
gg_cellphase(gg_cellphase_data = dm.cellphase, type = "facet")
}

}
\author{
Brent Thorne
}
