% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_cellphase_data.R
\name{gg_cellphase_data}
\alias{gg_cellphase_data}
\title{Create's a ggplot friendly data table for the dm.phase product.}
\usage{
gg_cellphase_data(dm.gpf, resolution = dendro.resolution(dm.gpf),shapeSensitivity = 0.6, minmaxDist = 0.2,minmaxSD = 2, radialIncrease = "max", na.omit = TRUE)
}
\arguments{
\item{dm.gpf}{a \code{data.frame} with either gap-free or gap-filled dendrometer series as produced by \code{\link[dendrometeR]{fill_gaps}}.}

\item{resolution}{a \code{numeric} specifying the resolution of the dendrometer data in seconds. Defaults to the resolution of \code{dm.gpf} as calculated using \code{\link[dendrometeR]{dendro.resolution}}.}

\item{shapeSensitivity}{a \code{numeric} specifying a time window, defined as proportion of a single day. Within this time window possible extrema points (i.e. minimum and maximum) in dendrometer measurements are searched for. Defaults to 0.6 (60\% of a day). See details for further explanation.}

\item{minmaxDist}{a \code{numeric} specifying the minimum temporal distance between consecutive minimum and maximum points (i.e. in the x direction). Defaults to 0.2 (20\% of a day).}

\item{minmaxSD}{a \code{numeric} specifying the minimum difference between consecutive minimum and maximum points expressed as a number of standard deviations (i.e. in the y direction). Defaults to 2.}

\item{radialIncrease}{a \code{character} string of \code{"max", "min", "mid"}, specifying when the stem-radius increment phase should start, with \code{"max"} as the most, and \code{"min"} as the least conservative approach; \code{"mid"} is in between. See details for further explanation.}

\item{na.omit}{a \code{logical opperator} statement of either \code{TRUE} or \code{FALSE}, specifying weather or not to remove \code{NA} values from final dataset. The default is set to \code{na.omit = TRUE}.}
}
\value{
The function returns a \code{data.frame} containing the following summary statistics:
\item{dmID}{dendrometer ID.}
\item{Phase}{cyclic phase (1: contraction, 2: expansion, 3: stem-radius increment, 4: full cycle).}
\item{timestamp}{timestamp indicating the date and time of phase stage.}
}
\description{
Transform the phase data into a single ggplot2 friendly data frame.
}
\examples{
\donttest{
library(dendrometeR)
data(dmCD)
dm.phase <- phase_def(dmCD)
dm.cellphase <- gg_cellphase_data(dmCD)
}

}
\author{
Brent Thorne
}
