% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggdendrometeR.R
\docType{package}
\name{ggdendrometeR}
\alias{ggdendrometeR}
\alias{ggdendrometeR-package}
\title{ggdendrometeR Package}
\description{
Add on to the dendrometeR package from CRAN that allows for more simple plotting with ggplot2
}
\details{
Includes functions:
\link{all_cycleStats}
\link{gg_cellphase}
}
\author{
Brent Thorne \email{brent.thorne18@gmail.com}
}
